# vox_api/docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Vox API server",
        "title": "Vox AI SI API",
        "contact": {
            "name": "Bogdan Antonovich",
            "email": "programjibogdan@gmail.com"
        },
        "version": "1.0"
    },
    "host": "smarttranslator.store/api",
    "basePath": "/",
    "paths": {
        "/admin/logs/level": {
            "put": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Set log level",
                "parameters": [
                    {
                        "description": "Log level (debug, info, warn, error)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/logs.LevelRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.HttpErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.HttpErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.HttpErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "Authenticates a user with the provided credentials, generates a JWT pair and sets them as cookies.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login with login and password",
                "parameters": [
                    {
                        "description": "Login payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.loginPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful, access and refresh token cookies set",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/models.HttpErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/models.HttpErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.HttpErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "description": "Validates the existing access and refresh token cookies, issues a new JWT pair and updates the cookies.",
                "tags": [
                    "auth"
                ],
                "summary": "Refresh JWT token pair",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access token cookie",
                        "name": "access_token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Refresh token cookie",
                        "name": "refresh_token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Token pair refreshed successfully, new cookies set"
                    },
                    "400": {
                        "description": "Access or refresh token cookie is missing",
                        "schema": {
                            "$ref": "#/definitions/models.HttpErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Refresh token is invalid",
                        "schema": {
                            "$ref": "#/definitions/models.HttpErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Access token is invalid",
                        "schema": {
                            "$ref": "#/definitions/models.HttpErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.HttpErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/sign_up": {
            "post": {
                "description": "Registers a new user with the provided credentials, generates a JWT pair and sets them as cookies.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Sign up with login and password",
                "parameters": [
                    {
                        "description": "Sign up payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.signUpPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User created successfully, access and refresh token cookies set"
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/models.HttpErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.HttpErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/{provider}/callback": {
            "get": {
                "description": "Handles the callback from the OAuth2 provider after user authorization. Exchanges the authorization code for a token, retrieves user info, creates the user if not exists, generates a JWT pair and sets them as cookies, then redirects to the frontend.",
                "tags": [
                    "auth"
                ],
                "summary": "OAuth2 provider callback",
                "parameters": [
                    {
                        "enum": [
                            "google",
                            "github"
                        ],
                        "type": "string",
                        "description": "OAuth2 provider",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization code returned by the provider",
                        "name": "code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "307": {
                        "description": "Redirect to the frontend with access and refresh token cookies set"
                    },
                    "401": {
                        "description": "Failed to exchange authorization code for token",
                        "schema": {
                            "$ref": "#/definitions/models.HttpErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Provider is not supported"
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.HttpErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/{provider}/login": {
            "get": {
                "description": "Redirects the user to the authorization page of the specified OAuth2 provider (google or github)",
                "tags": [
                    "auth"
                ],
                "summary": "Login via OAuth2 provider",
                "parameters": [
                    {
                        "enum": [
                            "google",
                            "github"
                        ],
                        "type": "string",
                        "description": "OAuth2 provider",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "307": {
                        "description": "Redirect to the provider authorization page"
                    },
                    "404": {
                        "description": "Provider is not supported"
                    }
                }
            }
        },
        "/hub/{hub_id}/listen": {
            "get": {
                "description": "Subscribes the client to a hub's audio stream and delivers synthesized audio chunks in real-time via chunked transfer encoding. The stream ends when the client disconnects or the hub closes.",
                "produces": [
                    "audio/mpeg"
                ],
                "tags": [
                    "hub"
                ],
                "summary": "Listen to audio stream",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hub ID",
                        "name": "hub_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Audio stream delivered as chunked audio/mpeg"
                    },
                    "404": {
                        "description": "Invalid hub ID (IsHubIDValid middleware)",
                        "schema": {
                            "$ref": "#/definitions/models.HttpErrorResponse"
                        }
                    }
                }
            }
        },
        "/hub/{hub_id}/new": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Creates a new hub with the given hub ID. Returns a conflict error if a hub with the same ID already exists.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hub"
                ],
                "summary": "Create a new hub",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hub ID",
                        "name": "hub_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Hub created successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Missing or invalid auth cookies (IsAuthorized middleware)",
                        "schema": {
                            "$ref": "#/definitions/models.HttpErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Hub ID is empty or invalid",
                        "schema": {
                            "$ref": "#/definitions/models.HttpErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Hub with the given ID already exists",
                        "schema": {
                            "$ref": "#/definitions/models.HttpErrorResponse"
                        }
                    },
                    "415": {
                        "description": "Invalid content type (IsContentTypeValid middleware)",
                        "schema": {
                            "$ref": "#/definitions/models.HttpErrorResponse"
                        }
                    }
                }
            }
        },
        "/hub/{hub_id}/publish": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Receives an audio stream, transcribes it via Deepgram, processes the transcription via Groq, and synthesizes speech via Fish Audio. All three operations run concurrently. Requires a valid user and hub in context.",
                "consumes": [
                    "application/octet-stream"
                ],
                "tags": [
                    "hub"
                ],
                "summary": "Publish audio stream",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hub ID",
                        "name": "hub_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Transcription language code (e.g. en, ru)",
                        "name": "lang",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Audio stream processed successfully"
                    },
                    "401": {
                        "description": "Missing or invalid auth cookies (IsAuthorized middleware)",
                        "schema": {
                            "$ref": "#/definitions/models.HttpErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Invalid user ID or hub",
                        "schema": {
                            "$ref": "#/definitions/models.HttpErrorResponse"
                        }
                    },
                    "415": {
                        "description": "Invalid content type (IsContentTypeValid middleware)",
                        "schema": {
                            "$ref": "#/definitions/models.HttpErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.HttpErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/info": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Returns the profile information of the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get user info",
                "responses": {
                    "200": {
                        "description": "User info returned successfully",
                        "schema": {
                            "$ref": "#/definitions/user.UserInfo"
                        }
                    },
                    "400": {
                        "description": "Invalid user ID",
                        "schema": {
                            "$ref": "#/definitions/models.HttpErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid auth cookies (IsAuthorized middleware)",
                        "schema": {
                            "$ref": "#/definitions/models.HttpErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.HttpErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/voice/new": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Receives a raw audio file (webm) as the request body and saves it as a voice reference for the authenticated user along with the provided reference text.",
                "consumes": [
                    "application/octet-stream"
                ],
                "tags": [
                    "voice"
                ],
                "summary": "Upload voice reference",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reference text associated with the voice sample",
                        "name": "text_ref",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Voice reference uploaded successfully"
                    },
                    "400": {
                        "description": "Reference text is missing",
                        "schema": {
                            "$ref": "#/definitions/models.HttpErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid auth cookies (IsAuthorized middleware)",
                        "schema": {
                            "$ref": "#/definitions/models.HttpErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Invalid user ID",
                        "schema": {
                            "$ref": "#/definitions/models.HttpErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.HttpErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "auth.loginPayload": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "auth.signUpPayload": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "login": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "logs.LevelRequest": {
            "type": "object",
            "properties": {
                "level": {
                    "type": "string"
                }
            }
        },
        "models.HttpErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "integer"
                        },
                        "message": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "user.UserInfo": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "picture": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "AdminAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "CookieAuth": {
            "type": "apiKey",
            "name": "access_token",
            "in": "cookie"
        },
        "CookieAuthRefresh": {
            "type": "apiKey",
            "name": "refresh_token",
            "in": "cookie"
        }
    }
}